+++ sdks/python/apache_beam/ml/inference/sklearn_inference_it_test.py
@@ -19,6 +19,7 @@
 
 import logging
 import re
+import sys
 import unittest
 import uuid
 
@@ -84,6 +85,8 @@ class SklearnInference(unittest.TestCase):
       true_label, expected_prediction = expected_outputs[i].split(',')
       self.assertEqual(predictions_dict[true_label], expected_prediction)
 
+  # TODO(https://github.com/apache/beam/issues/27151) use model with sklearn 1.2
+  @unittest.skipIf(sys.version_info >= (3, 11, 0), "Beam#27151")
   def test_sklearn_regression(self):
     test_pipeline = TestPipeline(is_integration_test=True)
     input_file = 'gs://apache-beam-ml/testing/inputs/japanese_housing_test_data.csv'  # pylint: disable=line-too-long